update dividends and governance

+++ core-contracts/Governance/src/test/java/network/balanced/score/core/governance/GovernanceTest.java
@@ -921,7 +921,7 @@ public class GovernanceTest extends GovernanceTestBase {
         governance.invoke(owner, "setRedeemBatchSize", _value);
 
         // Assert
-        verify(loans.mock).setRedeemBatchSize(_value);
+        verify(loans.mock).setRedeemBatchSize(_value.intValue());
 
     }
